Answered to the first commit comments
